Refaktor code and remove utils.py in samples
