+++ skopt/samples/base.py
@@ -1,0 +1,4 @@
+
+class InitialPointGenerator(object):
+    def generate(self, n_dim, n_samples, random_state=None):
+        raise NotImplemented
